+++ src/main/java/insane96mcp/progressivebosses/mixin/MixinStrafePlayerPhase.java
@@ -73,7 +73,8 @@ public abstract class MixinStrafePlayerPhase extends Phase {
 							this.dragon.getPhaseManager().setPhase(PhaseType.HOLDING_PATTERN);
 						//Otherwise reset the phase, in case she fireballs again
 						else
-							this.initPhase();
+							//Can't use initPhase() otherwise the target is reset
+							this.fireballCharge = 0;
 					}
 				}
 				else if (this.fireballCharge > 0) {